pyproject.toml
[project]
name = "mcpskill"
version = "0.2.0"
description = "Direct MCP Server Wrapper for Claude Code Skills"
requires-python = ">=3.10"
authors = [
    { name = "MCPSkill Contributors" }
]
readme = "README.md"
license = { text = "Apache-2.0" }

dependencies = [
    "mcp>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# No CLI scripts - this is a library for Claude Code skills

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mcpskill"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
